Complete overhaul of software database
Processes are now recorded by Names instead of executable names. It is more reliable and ergonomic
Before two process could have the same name for totally different softwares.
Each platform can have a different path for the executable
The detection is based of either the full path or the name of the binary
The UI has been changed to reflect all these changes